api/requirements.txt
fastapi
uvicorn
pandas
numpy
psycopg2-binary
requests
xgboost
joblib
python-dotenv
scikit-learn
